+++ apps/api/src/routes/skills.routes.ts
@@ -34,6 +34,7 @@ import { getUnifiedDatabase } from '../data/models/unified-database';
 import { getUnifiedMCPClient } from '../mcp/unified-client';
 import { requireAuth } from '../auth';
 import { assertResourceOwnerOrAdmin } from '../auth/ownership';
+import { exportSkill } from './skills-export';
 import { validate, validateQuery, validateWithSecurity } from '../middlewares/validation';
 import {
     createSkillSchema,
@@ -566,40 +567,8 @@ router.delete('/:skillId', requireAuth, asyncHandler(async (req: Request, res: R
 // 스킬 내보내기
 // ================================================
 
-/**
- * GET /api/agents/skills/:skillId/export
- * 스킬을 SKILL.md 파일로 내보내기
- */
-router.get('/:skillId/export', requireAuth, asyncHandler(async (req: Request, res: Response) => {
-    const { skillId } = req.params;
-    const userId = (req.user && 'userId' in req.user ? (req.user as { userId: string }).userId : req.user?.id?.toString());
-    const skill = await getSkillManager().getSkillById(skillId);
-    if (!skill) {
-        res.status(404).json(notFound('스킬'));
-        return;
-    }
-    // 형제 라우트(rewrite-proposal·PUT·DELETE)와 같은 소유권 가드 — getSkillById 는 id 만으로 조회하므로
-    // 비공개 타인 스킬의 본문(프롬프트 SoT)이 id 만 알면 내려가던 갭. 공개(isPublic) 스킬은 통과.
-    if (skill.createdBy && skill.isPublic !== true) {
-        assertResourceOwnerOrAdmin(String(skill.createdBy), String(userId), req.user?.role || 'user');
-    }
-
-    const markdown = [
-        `# ${skill.name}`,
-        '',
-        `> ${skill.description}`,
-        '',
-        `**Category**: ${skill.category}`,
-        '',
-        '## Instructions',
-        '',
-        skill.content
-    ].join('\n');
-
-    res.setHeader('Content-Type', 'text/markdown; charset=utf-8');
-    res.setHeader('Content-Disposition', `attachment; filename="${skill.name.replace(/[^a-z0-9_]/gi, '_').toLowerCase()}.SKILL.md"`);
-    res.send(markdown);
-}));
+/** GET /api/agents/skills/:skillId/export — 본문은 skills-export.ts (600줄 가드로 분리) */
+router.get('/:skillId/export', requireAuth, asyncHandler(exportSkill));
 
 export default router;
 export { router as skillsRouter };
